Fixed issue with set hole not on face of hex
Added code to put the set screw hole on the face of the hex.  Essentially added 30 degrees to the the position.  This could be simplified by adding the 30 degrees to the round position also.  However, it seemed that it would be better to keep the hex bore specific mods more separate.